backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.5.2
bcrypt==4.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
deep-translator==1.11.4
dnspython==2.7.0
easyocr==1.7.2
ecdsa==0.19.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.4
opencv-python-headless==4.11.0.86
packaging==24.2
passlib==1.7.4
pillow==11.2.1
pyasn1==0.4.8
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
python-bidi==0.6.6
python-dotenv==1.1.0
python-jose==3.4.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
sympy==1.13.1
tifffile==2025.3.30
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
